peres: cast data length to size_t
We expect that value in size_t and we're using uint*_t type for the
value in question.

Cast to size_t also eliminate -Wsign-compare when compile with gcc
on x86_64 platform.

+++ src/peres.c
@@ -223,7 +223,7 @@ static void peres_show_node(const pe_resource_node_t *node)
 			output("String len", value);
 
 			char ascii_string[MAX_MSG];
-			size_t min_size = pe_utils_min(sizeof(ascii_string), dataString->Length + 1);
+			size_t min_size = pe_utils_min(sizeof(ascii_string), (size_t)dataString->Length + 1);
 			pe_utils_str_widechar2ascii(ascii_string, (const char *)dataString->String, min_size);
 			ascii_string[min_size - 1] = '\0'; // Null terminate it.
